extract common functions from v1/v2 designer panels
